PWC-159: File Deletion (#102)

+++ src/hooks/firebase/useArticle.ts
@@ -17,17 +17,75 @@ import {
   getDoc,
   getDocs,
   startAfter,
+  Timestamp,
 } from 'firebase/firestore'
-import { getDownloadURL, ref, uploadBytes } from '@firebase/storage'
 import { useState, useEffect, useContext } from 'react'
 import { db, storage } from '../../firebaseApp'
 import { useAuth } from './useAuth'
 import { NotificationContext } from '../../context/NotificationContext'
-import { Article, ArticlePreview } from 'types/Article'
-import { articlePreviewTranslator } from 'utils/firebase/article'
-import { UserData } from 'types/UserData'
+import {deleteObject, getDownloadURL, ref, StorageError, uploadBytes} from "@firebase/storage";
 import { UserComment } from 'types/Comment'
+import { UserData } from 'types/UserData'
 import { getUser } from 'utils/firebase/user'
+
+export interface ArticlePreview {
+  title: string
+  content: string
+  header_image: string
+  author_image: string
+  author_username: string
+  publish_time: Timestamp
+  articleId: string
+}
+
+export const articlePreviewTranslator = (
+  docs: QuerySnapshot<DocumentData>,
+): ArticlePreview[] => {
+  const articlesData: ArticlePreview[] = []
+  docs.forEach((doc) => {
+    articlesData.push({
+      title: doc.data().title,
+      content: doc.data().content,
+      header_image: doc.data().header_image,
+      author_image: doc.data().author_image,
+      author_username: doc.data().author_username,
+      publish_time: doc.data().publish_time,
+      articleId: doc.id,
+    })
+  })
+
+  return articlesData
+}
+
+export const articleDraftTranslator = (
+  docs: QuerySnapshot<DocumentData>,
+): ArticlePreview[] => {
+  const articlesData: ArticlePreview[] = []
+  docs.forEach((doc) => {
+    articlesData.push({
+      title: doc.data().title,
+      content: doc.data().content,
+      header_image: doc.data().header_image,
+      author_image: doc.data().author_image,
+      author_username: doc.data().author_username,
+      publish_time: doc.data().edit_time,
+      articleId: doc.id,
+    })
+  })
+
+  return articlesData
+}
+
+export interface Article {
+  title: string
+  content: string
+  header_image: string
+  author_image: string
+  author_uid: string
+  edit_time: Timestamp
+  author_username: string
+  publish_time: Timestamp
+}
 
 export const useArticleRecents = (n: number) => {
   const [error, setError] = useState<FirestoreErrorCode>()
@@ -175,41 +233,54 @@ export const useUploadHeader = () => {
   // eslint-disable-next-line @typescript-eslint/no-var-requires
   const uuid = require('uuid')
 
-  const uploadHeader = async (file: File) => {
-    if (currentUser == null) {
-      setError('unauthenticated')
-      return
-    }
-    setLoading(true)
-    const path = `${currentUser.uid}/${file.name}${uuid.v4()}`
-    // const storageRef = ref(storage, `${currentUser.uid}/${file.name}${uuid.v4()}`)
-    // try {
-    //     await uploadBytes(storageRef, file)
-    //     const url = await getDownloadURL(storageRef)
-    //     setImageURL(url)
-    //     setLoading(false)
-    // } catch (err) {
-    //     if (err instanceof FirestoreError) {
-    //         setError(err.code)
-    //     } else {
-    //         setError("unknown-error")
-    //     }
-    //     setLoading(false)
-    const storageRef = ref(storage, path)
-    uploadBytes(storageRef, file)
-      .then(() => {
-        getDownloadURL(storageRef).then((res) => {
-          setLoading(false)
-          setImageURL(res)
+    const uploadHeader = async (file: File) => {
+        if(currentUser == null) {
+            setError("unauthenticated");
+            return
+        }
+        setLoading(true)
+        const path = `${currentUser.uid}/${file.name}${uuid.v4()}`
+        const storageRef = ref(storage, path)
+        uploadBytes(storageRef, file).then(() => {
+            getDownloadURL(storageRef).then((res) => {
+                setLoading(false)
+                setImageURL(res)
+            })
+        }).catch((err) => {
+            setLoading(false)
+            setError(err.code)
         })
-      })
-      .catch((err) => {
-        setLoading(false)
-        setError(err.code)
-      })
-  }
+        
+    }
+  
 
   return { uploadHeader, error, loading, imageURL }
+}
+
+const deleteStorage = (imageURL: string): Promise<void> => {
+    try{
+        const reference = ref(storage, imageURL)
+        return deleteObject(reference)
+    }
+    catch (e) {
+        return Promise.resolve(); // Not a storage image, no behaviour required
+    }
+}
+
+export const useDeleteHeader = () => {
+    const [error, setError] = useState<string>()
+    const [loading, setLoading] = useState(false)
+
+    const deleteHeader = (imageURL: string) => {
+        setLoading(true)
+        deleteStorage(imageURL).then(() => {
+            setLoading(false)
+        }).catch((err)=> {
+            setError(err.code)
+        })
+    }
+
+    return {deleteHeader, error, loading}
 }
 
 export const useArticleCreate = () => {
@@ -291,25 +362,40 @@ export const useArticleEdit = () => {
 export const useArticleDelete = (articleID: string) => {
   const { dispatch } = useContext(NotificationContext)
   const [error, setError] = useState<FirestoreErrorCode>()
-  const [loading, setLoading] = useState(true)
+  const [loading, setLoading] = useState(false)
 
-  const deleteArticle = async () =>
-    deleteDoc(doc(db, 'article', articleID)).then(
-      () => {
-        dispatch({
-          notificationActionType: 'success',
-          message: `Successfuly deleted article`,
-        })
-        setLoading(false)
-      },
-      (err) => {
-        dispatch({
-          notificationActionType: 'error',
-          message: `Error deleting article. Error code: ${err.code}`,
-        })
-        setError(err.code)
-      },
-    )
+  const deleteArticle = async () => {
+      getDoc(doc(db, 'article', articleID))
+          .then((document) => {
+                  const data = document.data()
+                  if (data === undefined) {
+                      setError('not-found')
+                      return
+                  }
+                  deleteStorage(data.header_image).catch((err) => setError(err.code))
+                  deleteDoc(doc(db, 'article', articleID)).then(
+                      () => {
+                          dispatch({
+                              notificationActionType: 'success',
+                              message: `Successfuly deleted article`,
+                          })
+                          setLoading(false)
+                      },
+                      (err) => {
+                          dispatch({
+                              notificationActionType: 'error',
+                              message: `Error deleting article. Error code: ${err.code}`,
+                          })
+                          setError(err.code)
+                      },
+                  )
+              }
+          ).catch((err) => {
+              setError(err.code)
+      })
+  }
+
+
 
   return { deleteArticle, error, loading }
 }
